requirements.txt
PyJWT==2.10.1
asgiref==3.11.0
black==25.11.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
defusedxml==0.7.1
django==5.2.8
django-cleanup==9.0.0
django-cors-headers==4.9.0
django-debug-toolbar==6.1.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
djoser==2.3.3
drf-spectacular==0.29.0
idna==3.11
inflection==0.5.1
isort==7.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mypy-extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==12.0.0
pip==25.3
platformdirs==4.5.0
psycopg2-binary==2.9.11
pycparser==2.23
python3-openid==3.2.0
pytokens==0.3.0
pyyaml==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.29.0
setuptools==65.5.0
sqlparse==0.5.4
typing-extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
